Timer Pause
Added Pause and Resume functions to Timer
