+++ HiarethTimer/Assets/Canvas.prefab
@@ -259,7 +259,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 50.6, y: 426}
+  m_AnchoredPosition: {x: 50.6, y: 398.7}
   m_SizeDelta: {x: 160, y: 18.9521}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &3508937867809795888
