+++ HiarethTimer/Assets/Scenes/SampleScene.unity
@@ -294,6 +294,17 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 2
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1787846114 stripped
+MonoBehaviour:
+  m_CorrespondingSourceObject: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+  m_PrefabInstance: {fileID: 3508937866729475045}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fe87c0e1cc204ed48ad3b37840f39efc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &3508937866729475045
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -301,6 +312,10 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 3508937867121954723, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Value
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 3508937867362233849, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
       propertyPath: m_Pivot.x
       value: 0
@@ -389,6 +404,66 @@ PrefabInstance:
       propertyPath: m_Name
       value: Canvas
       objectReference: {fileID: 0}
+    - target: {fileID: 3508937867809795891, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Pivot.x
+      value: 0.52353406
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937867809795891, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Pivot.y
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937867809795891, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 17.7221
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937867809795891, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 53.81655
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937867809795891, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 398.8
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Type
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Color.b
+      value: 0.5673151
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Color.g
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Color.r
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_Maskable
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_FillAmount
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_FillCenter
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_FillMethod
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3508937868638646824, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
+      propertyPath: m_FillOrigin
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 3508937868638646825, guid: 204ca06c669c8de40a4e4ffa93598cad, type: 3}
       propertyPath: m_AnchorMax.x
       value: 0
@@ -406,6 +481,46 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: myColor.b
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: myColor.g
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: myColor.r
+      value: 0.9901844
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: fullTime.b
+      value: 0.78488207
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: fullTime.g
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: fullTime.r
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: halfTime.b
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: halfTime.g
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: halfTime.r
+      value: 0.95075846
+      objectReference: {fileID: 0}
+    - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
+      propertyPath: sliderFill
+      value: 
+      objectReference: {fileID: 1787846114}
     - target: {fileID: 9170270669868353752, guid: 88a156f87e0651f4f983efaec931967d, type: 3}
       propertyPath: timeDisplay
       value:
