+++ HiarethTimer/Assets/Scripts/TimerSlider.cs
@@ -1,0 +1,89 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class TimerSlider : MonoBehaviour
+{
+    public Image sliderFill;
+    public Color myColor;
+    public Slider timerSlider;
+    public Text timeDisplay;
+    public float startTime;
+    float gameTime;
+    float timeScaleValue;
+    
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        timerSlider.maxValue = startTime;
+        timerSlider.value = gameTime;
+        myColor.a = 1;
+    }
+
+    // Update is called once per frame
+    void Update(){
+        timeUpdate();  
+            if(gameTime <=0){
+                gameTime = startTime;
+                sliderUpdate();
+            }
+            else
+                sliderUpdate();     
+    }
+
+     void sliderUpdate(){
+        int minutes = Mathf.FloorToInt(gameTime / 60);
+        int seconds = Mathf.FloorToInt(gameTime - minutes * 60f);
+        string textTime = string.Format("{0:0}:{1:00}", minutes, seconds);  
+        timeDisplay.text = textTime;
+        timerSlider.value = gameTime;
+    }
+
+    void timeUpdate(){
+        gameTime -= Time.deltaTime;
+    }
+
+
+    public void pauseTime(){
+        timeScaleValue = Time.timeScale;
+        Time.timeScale = 0;
+        sliderUpdate();             
+     }
+
+     public void resumeTime(){
+        if(timeScaleValue == 2)
+            doubleTime();
+        else
+            normalTime();        
+     }  
+
+    public void doubleTime(){
+        Time.timeScale = 2;
+        sliderUpdate();             
+     } 
+
+    public void normalTime(){
+        Time.timeScale = 1;
+        sliderUpdate();             
+     }  
+
+    public void removeTime(){
+        gameTime = gameTime - 2;
+        sliderUpdate();             
+     } 
+
+    public void addTime(){
+        gameTime = gameTime + 2;
+        sliderUpdate();             
+     }    
+
+     void colorUpdate(){
+        if(gameTime <=10)
+            sliderFill.color = myColor;
+        }
+
+
+   
+}
