--- HiarethTimer/Assets/TimerScript/TimerSlider.cs
@@ -1,68 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class TimerSlider : MonoBehaviour
-{
-    public Slider timerSlider;
-    public Text timeDisplay;
-    float gameTime;
-    public float startTime;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        timerSlider.maxValue = startTime;
-        timerSlider.value = gameTime;
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        timeUpdate();  
-
-        if(gameTime <=0){
-            gameTime = startTime;
-            sliderUpdate();
-        }
-        else{
-           sliderUpdate();
-        }
-           
-    }
-
-
-    public void doubleTime(){
-        Time.timeScale = 2;
-        sliderUpdate();             
-     } 
-
-    public void normalTime(){
-        Time.timeScale = 1;
-        sliderUpdate();             
-     }  
-
-    public void removeTime(){
-        gameTime = gameTime - 2;
-        sliderUpdate();             
-     } 
-
-    public void addTime(){
-        gameTime = gameTime + 2;
-        sliderUpdate();             
-     }    
-
-
-    void sliderUpdate(){
-        int minutes = Mathf.FloorToInt(gameTime / 60);
-        int seconds = Mathf.FloorToInt(gameTime - minutes * 60f);
-        string textTime = string.Format("{0:0}:{1:00}", minutes, seconds);  
-        timeDisplay.text = textTime;
-        timerSlider.value = gameTime;
-    }
-
-    void timeUpdate(){
-        gameTime -= Time.deltaTime;
-    }
-}
